refact: ajuste nas entidades, seed e properties

+++ src/main/java/com/sys/gestaoatividades/domain/repository/ParticipanteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticipanteRepository extends JpaRepository<Participante, Long> {
+public interface ParticipanteRepository extends JpaRepository<Participante, Integer> {
 }
